Allow the EV3 to hang on exit like other nerves systems
The EV3 system does not have a well known reboot button so many people
pull the battery to reboot the EV3. In favor of not losing important
debugging information it would be nice to not have the brick poweroff
when it hangs, and to consistently behave across nerves systems.

In order to make it simpler I documented the reboot process of the EV3
in the README. If the user would like to go back to the old behaviour I
also included directions to change the `erlinit.config`

closes !9

Amos King @adkron <amos@binarynoggin.com>